requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
joblib
matplotlib
mlflow>=2.0
dvc>=2.0
psycopg2-binary
python-dotenv
pytest
pydantic
seaborn
